+++ src/Yaat.Sim/Commands/CanonicalCommandType.cs
@@ -268,7 +268,6 @@ public enum CanonicalCommandType
     HalfStripMove,
     HalfStripOffset,
     HalfStripSlide,
-    HalfStripEdit,
     SeparatorCreate,
     SeparatorDelete,
     SeparatorEdit,