+++ src/Yaat.Sim/Commands/TrackEngine.cs
@@ -72,7 +72,6 @@ public static class TrackEngine
                 or HalfStripMoveCommand
                 or HalfStripOffsetCommand
                 or HalfStripSlideCommand
-                or HalfStripEditCommand
                 or SeparatorCreateCommand
                 or SeparatorDeleteCommand
                 or SeparatorEditCommand